Fix _CDB_UserQuotaInBytes return to int8
bytes quota often exceeds int4...

+++ lib/sql/scripts-available/CDB_CartodbfyTable.sql
@@ -36,7 +36,7 @@ DECLARE
   had_column BOOLEAN;
   i INTEGER;
   new_name TEXT;
-  quota_in_bytes INT;
+  quota_in_bytes INT8;
 BEGIN
 
   -- Ensure required fields exist
